[GBA] Another attempt to fix ARM ADC and SBC

+++ src/gba/arm7.h
@@ -257,7 +257,7 @@ class ARM7
 
 	//Misc CPU helpers
 	void update_condition_logical(u32 result, u8 shift_out);
-	void update_condition_arithmetic(u32 input, u32 operand, u32 result, bool addition);
+	void update_condition_arithmetic(u32 input, u64 operand, u32 result, bool addition);
 	bool check_condition(u32 current_arm_instruction) const;
 	u8 logical_shift_left(u32& input, u8 offset);
 	u8 logical_shift_right(u32& input, u8 offset);
